# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GradeFlow API
servers:
  - url: /api
    description: Base API path
security:
  - cookieAuth: []
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      responses:
        "200":
          description: All grader accounts (admin only)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
        "403":
          description: Forbidden
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    post:
      operationId: createUser
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUserRequest" }
      responses:
        "201":
          description: User created (admin only)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreatedUserResponse" }
        "400":
          description: Email already in use or invalid input
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /users/{userId}:
    parameters:
      - in: path
        name: userId
        required: true
        schema: { type: integer }
    patch:
      operationId: updateUser
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateUserRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
    delete:
      operationId: deleteUser
      tags: [users]
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
        "400":
          description: Cannot delete the current account
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /users/{userId}/reset-password:
    parameters:
      - in: path
        name: userId
        required: true
        schema: { type: integer }
    post:
      operationId: adminResetUserPassword
      tags: [users]
      responses:
        "200":
          description: |
            Generates a temporary password for the user and revokes all of
            their existing sessions. The temporary password is returned to
            the admin so they can hand it off out-of-band.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminResetPasswordResponse" }
...
  /assignments:
    get:
      operationId: listAssignments
      tags: [assignments]
      parameters:
        - in: query
          name: courseId
          schema: { type: integer }
      responses:
        "200":
          description: Assignments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Assignment" }
    post:
      operationId: createAssignment
      tags: [assignments]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AssignmentInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Assignment" }
...
  /assignments/{assignmentId}/submissions:
    parameters:
      - $ref: "#/components/parameters/AssignmentId"
    patch:
      operationId: bulkUpdateSubmissions
      tags: [assignments]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BulkSubmissionUpdate" }
      responses:
        "200":
          description: Bulk update result
          content:
            application/json:
              schema:
                type: object
                required: [updated]
                properties:
                  updated: { type: integer }
    get:
      operationId: listSubmissions
      tags: [submissions]
      responses:
        "200":
          description: Submissions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SubmissionWithStudent" }
  /submissions/{submissionId}:
    parameters:
      - $ref: "#/components/parameters/SubmissionId"
    get:
      operationId: getSubmission
      tags: [submissions]
      responses:
        "200":
          description: Submission
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubmissionDetail" }
    patch:
      operationId: updateSubmission
      tags: [submissions]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubmissionInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubmissionDetail" }
  /submissions/{submissionId}/file:
    parameters:
      - $ref: "#/components/parameters/SubmissionId"
    post:
      operationId: setSubmissionFile
      tags: [submissions]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubmissionFileInput" }
      responses:
        "200":
          description: File set
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubmissionDetail" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: gradeflow_token
      description: |
        Session JWT cookie set by `POST /auth/login`. All endpoints require it
        unless the operation explicitly overrides `security: []`.
  parameters:
...
    AssignmentId:
      in: path
      name: assignmentId
      required: true
      schema: { type: integer }
    SubmissionId:
      in: path
      name: submissionId
      required: true
      schema: { type: integer }
  schemas:
...
    Ok:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    UserRole:
      type: string
      enum: [admin, lecturer, grader]
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
        role: { $ref: "#/components/schemas/UserRole" }
        createdAt: { type: string, format: date-time }
    CreateUserRequest:
      type: object
      required: [email, name, role]
      properties:
        email: { type: string, format: email }
        name: { type: string, minLength: 1 }
        role: { $ref: "#/components/schemas/UserRole" }
        password:
          type: string
          minLength: 8
          nullable: true
          description: |
            Optional. When omitted, the server generates a random temporary
            password and returns it in the response so the admin can share it
            with the new user.
    CreatedUserResponse:
      type: object
      required: [user]
      properties:
        user: { $ref: "#/components/schemas/User" }
        generatedPassword:
          type: string
          nullable: true
          description: |
            Random password generated when the request omitted `password`.
            Shown only once — admins must save or share it immediately.
    UpdateUserRequest:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        role: { $ref: "#/components/schemas/UserRole" }
    AdminResetPasswordResponse:
      type: object
      required: [temporaryPassword]
      properties:
        temporaryPassword: { type: string }
...
    Student:
      type: object
      required: [id, externalId, firstName, lastName]
      properties:
        id: { type: integer }
        externalId: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        notes: { type: string, nullable: true }
...
    Assignment:
      type: object
      required: [id, courseId, name, maxScore, weight, gradingScale, closed]
      properties:
        id: { type: integer }
        courseId: { type: integer }
        courseName: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        dueDate: { type: string, format: date-time, nullable: true }
        maxScore: { type: number }
        weight: { type: number }
        gradingScale: { type: string }
        closed: { type: boolean }
        submissionCount: { type: integer }
        gradedCount: { type: integer }
...
    AssignmentInput:
      type: object
      required: [courseId, name, maxScore, weight, gradingScale]
      properties:
        courseId: { type: integer }
        name: { type: string, minLength: 1 }
        description: { type: string, nullable: true }
        dueDate: { type: string, format: date-time, nullable: true }
        maxScore: { type: number }
        weight: { type: number }
        gradingScale: { type: string }
...
    Submission:
      type: object
      required: [id, assignmentId, studentId, status, originalityFlag, submittedLate]
      properties:
        id: { type: integer }
        assignmentId: { type: integer }
        studentId: { type: integer }
        status:
          type: string
          enum: [pending, in_progress, needs_review, graded, returned, missing]
        score: { type: number, nullable: true }
        feedback: { type: string, nullable: true }
        privateNotes: { type: string, nullable: true }
        originalityFlag: { type: boolean }
        submittedLate: { type: boolean }
        submittedAt: { type: string, format: date-time, nullable: true }
        fileObjectPath: { type: string, nullable: true }
        fileName: { type: string, nullable: true }
        contentType: { type: string, nullable: true }
        fileSize: { type: integer, nullable: true }
        gradedAt: { type: string, format: date-time, nullable: true }
        updatedAt: { type: string, format: date-time, nullable: true }
    SubmissionWithStudent:
      allOf:
        - $ref: "#/components/schemas/Submission"
        - type: object
          properties:
            student: { $ref: "#/components/schemas/Student" }
...
    SubmissionDetail:
      allOf:
        - $ref: "#/components/schemas/SubmissionWithStudent"
        - type: object
          properties:
            assignment: { $ref: "#/components/schemas/Assignment" }
            rubricScores:
              type: array
              items: { $ref: "#/components/schemas/RubricScore" }
    BulkSubmissionUpdate:
      type: object
      required: [submissionIds, status]
      properties:
        submissionIds:
          type: array
          minItems: 1
          items: { type: integer }
        status:
          type: string
          enum: [pending, in_progress, needs_review, graded, returned, missing]
    SubmissionInput:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in_progress, needs_review, graded, returned, missing]
        score: { type: number, nullable: true }
        feedback: { type: string, nullable: true }
        privateNotes: { type: string, nullable: true }
        originalityFlag: { type: boolean }
        submittedLate: { type: boolean }
        submittedAt: { type: string, format: date-time, nullable: true }
    SubmissionFileInput:
      type: object
      required: [objectPath, fileName, contentType, fileSize]
      properties:
        objectPath: { type: string }
        fileName: { type: string }
        contentType: { type: string }
        fileSize: { type: integer }
...
    RubricScore:
      type: object
      required: [id, submissionId, criterionId, points]
      properties:
        id: { type: integer }
        submissionId: { type: integer }
        criterionId: { type: integer }
        points: { type: number }
        comment: { type: string, nullable: true }
...